pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["Daria Zaichikova"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
django-environ = "^0.9.0"
pillow = "^9.3.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
social-auth-app-django = " ^5.0.0"
django-filter = "^22.1"
marshmallow-dataclass = "^8.5.11"
requests = "^2.28.2"
pydantic = "^1.10.4"
pytest = "^7.2.1"
cffi = "^1.15.1"
ansible-vault = "^2.1.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"
mypy = "^1.0.0"


[tool.poetry.group.tests.dependencies]
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
